refactor(sdk): separate pure agent API from subscription logic
