+++ libs/sdk/src/api/index.ts
@@ -1,4 +1,5 @@
 export * from "./agent";
+export * from "./agent-subscription";
 export * from "./app";
 export * from "./auth";
 export * from "./base";
